+++ include/mruby_integration/core/timing.hpp
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "mruby.h"
-
-void
-append_core_timing(mrb_state*);