+++ mrb_doc/core/timing.rb
@@ -1,7 +1,0 @@
-# Returns the amount of time the has passed since {Window.open} was called.
-# @return [Float]
-def get_time
-  # mrb_time
-  # src/mruby_integration/core/timing.cpp
-  36.345
-end