+++ src/mruby_integration/core.cpp
@@ -2,7 +2,6 @@
 
 #include "mruby_integration/core/drawing.hpp"
 #include "mruby_integration/core/misc.hpp"
-#include "mruby_integration/core/timing.hpp"
 
 #include "ruby/core/buildkite_analytics.hpp"
 
@@ -11,7 +10,6 @@ append_core(mrb_state* mrb)
 {
   append_core_drawing(mrb);
   append_core_misc(mrb);
-  append_core_timing(mrb);
 
   load_ruby_core_buildkite_analytics(mrb);
 }