+++ src/mruby_integration/models/window.cpp
@@ -438,6 +438,12 @@ mrb_Window_frame_time(mrb_state* mrb, mrb_value) -> mrb_value
   return mrb_float_value(mrb, GetFrameTime());
 }
 
+auto
+mrb_Window_seconds_open(mrb_state* mrb, mrb_value) -> mrb_value
+{
+  return mrb_float_value(mrb, GetTime());
+}
+
 void
 append_models_Window(mrb_state* mrb)
 {
@@ -523,6 +529,11 @@ append_models_Window(mrb_state* mrb)
     mrb, Window_class, "frame_rate", mrb_Window_frame_rate, MRB_ARGS_NONE());
   mrb_define_class_method(
     mrb, Window_class, "frame_time", mrb_Window_frame_time, MRB_ARGS_NONE());
+  mrb_define_class_method(mrb,
+                          Window_class,
+                          "seconds_open",
+                          mrb_Window_seconds_open,
+                          MRB_ARGS_NONE());
 
   load_ruby_models_window(mrb);
 }